Support MacOS
This (surprisingly small) commit adds functional and CI support for
MacOS. The main blocker on supporting MacOS was a runtime linking error
caused by libaccp trying to load a `libcrypto.so` it found in some
`/usr` system directory on MacOS, instead of the temporary directory
that ACCP creates and restricts `rpath` to during linking. The reason
for this is that the `rpath` value of `$ORIGIN` is specific to linux's
linker, and is meaningless on MacOS. So, the MacOS linker consulted a
non-existent `rpath`, didn't find anything, and fell back to system
library directories. The fix is to detect OS type at build time and use
the appropriate value for `rpath` (on MacOS that's `@loader_path`, see
the "Locating External Resources" section of [this documentation
page][1]). We also explicity set `CMAKE_MACOSX_RPATH` on MacOS, which is
documented [here][2].

The other pieces of this commit are related to CI. In order for the
coverate steps to pass, we need to install `lcov` and `gcovr`
dependencies on the GitHub Actions MacOS CI images. Lastly, now that
this runs on Mac the Apple JCE providers in our pre-baked properties
files are actually used (they're ignored on Linux, where they're
unavailable). This resulted in the same provider being installed twice,
causing SecurityPropertiesTester failures. We also canonicalize provider
names wth fully qualified class names in our duplicate install check
(read through docs [here][3] and [here][4] for more information about
registering JCE providers).

This commit resolves [Issue #41][5].

[1]: https://developer.apple.com/library/archive/documentation/DeveloperTools/Conceptual/DynamicLibraries/100-Articles/DynamicLibraryDesignGuidelines.html#//apple_ref/doc/uid/TP40002013-SW21
[2]: https://cmake.org/cmake/help/latest/prop_tgt/MACOSX_RPATH.html
[3]: https://docs.oracle.com/javase/8/docs/api/java/security/Provider.html
[4]: https://docs.oracle.com/javase/8/docs/technotes/guides/security/crypto/CryptoSpec.html#Provider
[5]: https://github.com/corretto/amazon-corretto-crypto-provider/issues/41

chkpnt: Add cmake rpath check, fix provider numbering, revert property checker class canonicalization

fix for linux

try while ready test